Fixed warning.
git-svn-id: file:///home/nick/svnrepo/repo/indigo/trunk@15116 20aa11e4-6e6f-594c-a032-6291ace9ee65

+++ utils/CryptoRNG.cpp
@@ -64,7 +64,7 @@ void getRandomBytes(uint8* buf, size_t buflen)
 		buflen, // buflen
 		0 // flags
 	);
-	if(res < buflen)
+	if(res < (ssize_t)buflen)
 		throw glare::Exception("getrandom failed: " + PlatformUtils::getLastErrorString());
 
 #endif
